+++ integrations/github/github/clients/utils.py
@@ -1,5 +1,6 @@
-from typing import Any, Dict, cast, TYPE_CHECKING
-from github.core.options import ListOrganizationOptions
+from typing import Any, Dict, Optional, cast, TYPE_CHECKING
+from github.core.options import ListOrganizationOptions, ListRepositoryOptions
+from github.helpers.models import RepoSearchParams
 from port_ocean.context.ocean import ocean
 
 from github.clients.auth.abstract_authenticator import AbstractGitHubAuthenticator
@@ -24,9 +25,28 @@ def get_github_organizations() -> ListOrganizationOptions:
     return ListOrganizationOptions(
         organization=organization,
         allowed_multi_organizations=port_app_config.organizations,
+        allow_personal_organization=port_app_config.allow_personal_organization,
     )
 
 
 def get_mono_repo_organization(organization: str | None) -> str | None:
     """Get the organization for a monorepo."""
     return organization or ocean.integration_config["github_organization"]
+
+
+def get_repository_options(
+    organization: str,
+    organization_type: str,
+    search_params: Optional[RepoSearchParams] = None,
+    included_relationships: Optional[list[str]] = None,
+) -> ListRepositoryOptions:
+
+    port_app_config = cast("GithubPortAppConfig", event.port_app_config)
+
+    return ListRepositoryOptions(
+        organization=organization,
+        organization_type=organization_type,
+        type=port_app_config.repository_type,
+        search_params=search_params,
+        included_relationships=included_relationships,
+    )