+++ integrations/github/integration.py
@@ -239,6 +239,11 @@ class GithubPortAppConfig(PortAppConfig):
             "member of) for Classic PAT authentication."
         ),
     )
+    allow_personal_organization: bool = Field(
+        default=False,
+        alias="allowPersonalOrganization",
+        description="Allow the personal organization to be synced",
+    )
     repository_type: str = Field(alias="repositoryType", default="all")
     resources: list[
         GithubRepositoryConfig