+++ integrations/github/main.py
@@ -19,7 +19,11 @@ from github.clients.client_factory import (
     create_github_client,
 )
 from github.core.exporters.workflow_runs_exporter import RestWorkflowRunExporter
-from github.clients.utils import get_github_organizations, integration_config
+from github.clients.utils import (
+    get_github_organizations,
+    get_repository_options,
+    integration_config,
+)
 from github.core.exporters.abstract_exporter import AbstractGithubExporter
 from github.core.exporters.branch_exporter import RestBranchExporter
 from github.core.exporters.deployment_exporter import RestDeploymentExporter
@@ -53,7 +57,6 @@ from github.core.options import (
     ListEnvironmentsOptions,
     ListIssueOptions,
     ListPullRequestOptions,
-    ListRepositoryOptions,
     ListTeamOptions,
     ListUserOptions,
     ListWorkflowOptions,
@@ -65,7 +68,11 @@ from github.core.options import (
     ListCollaboratorOptions,
     ListSecretScanningAlertOptions,
 )
-from github.helpers.utils import ObjectKind, GithubClientType
+from github.helpers.utils import (
+    ObjectKind,
+    GithubClientType,
+    enrich_user_with_primary_email,
+)
 from github.webhook.events import WEBHOOK_CREATE_EVENTS
 from github.webhook.webhook_client import GithubWebhookClient
 
@@ -120,9 +127,6 @@ async def on_start() -> None:
 
     github_organization = ocean.integration_config.get("github_organization")
     if github_organization:
-        logger.info(
-            f"Subscribing to GitHub webhooks for organization: {github_organization}"
-        )
         await _create_webhooks_for_organization(github_organization, base_url)
         return
 
@@ -161,21 +165,19 @@ async def resync_repositories(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
 
     rest_client = create_github_client()
     org_exporter = RestOrganizationExporter(rest_client)
-    port_app_config = cast(GithubPortAppConfig, event.port_app_config)
-
     repo_config = cast(GithubRepositoryConfig, event.resource_config)
-    included_relationships = repo_config.selector.include
+    included_relationships = cast(list[str], repo_config.selector.include)
 
     async for organizations in org_exporter.get_paginated_resources(
         get_github_organizations()
     ):
         tasks = (
             RestRepositoryExporter(rest_client).get_paginated_resources(
-                options=ListRepositoryOptions(
-                    organization=org["login"],
-                    type=port_app_config.repository_type,
-                    included_relationships=cast(list[str], included_relationships),
-                    search_params=repo_config.selector.repo_search,
+                options=get_repository_options(
+                    org["login"],
+                    org["type"],
+                    repo_config.selector.repo_search,
+                    included_relationships,
                 )
             )
             for org in organizations
@@ -194,20 +196,30 @@ async def resync_users(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     org_exporter = RestOrganizationExporter(rest_client)
     user_config = cast(GithubUserConfig, event.resource_config)
     include_bots = user_config.selector.include_bots
+    exporter = GraphQLUserExporter(graphql_client)
 
     async for organizations in org_exporter.get_paginated_resources(
         get_github_organizations()
     ):
-        tasks = (
-            GraphQLUserExporter(graphql_client).get_paginated_resources(
-                options=ListUserOptions(
-                    organization=org["login"], include_bots=include_bots
+        tasks = []
+        for org in organizations:
+            if org["type"] == "Organization":
+                tasks.append(
+                    exporter.get_paginated_resources(
+                        options=ListUserOptions(
+                            organization=org["login"], include_bots=include_bots
+                        )
+                    )
                 )
-            )
-            for org in organizations
-        )
-        async for users in stream_async_iterators_tasks(*tasks):
-            yield users
+                continue
+
+            if not org.get("email"):
+                org = await enrich_user_with_primary_email(rest_client, org)
+            yield [org]
+
+        if tasks:
+            async for users in stream_async_iterators_tasks(*tasks):
+                yield users
 
 
 @ocean.on_resync(ObjectKind.TEAM)
@@ -228,20 +240,24 @@ async def resync_teams(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     ):
         tasks = []
         for org in organizations:
-            org_name = org["login"]
-            exporter: AbstractGithubExporter[Any]
+            if org["type"] == "Organization":
+                org_name = org["login"]
+                exporter: AbstractGithubExporter[Any]
 
-            if selector.members:
-                exporter = GraphQLTeamWithMembersExporter(graphql_client)
-            else:
-                exporter = RestTeamExporter(rest_client)
+                if selector.members:
+                    exporter = GraphQLTeamWithMembersExporter(graphql_client)
+                else:
+                    exporter = RestTeamExporter(rest_client)
 
-            tasks.append(
-                exporter.get_paginated_resources(ListTeamOptions(organization=org_name))
-            )
+                tasks.append(
+                    exporter.get_paginated_resources(
+                        ListTeamOptions(organization=org_name)
+                    )
+                )
 
-        async for teams in stream_async_iterators_tasks(*tasks):
-            yield teams
+        if tasks:
+            async for teams in stream_async_iterators_tasks(*tasks):
+                yield teams
 
 
 @ocean.on_resync(ObjectKind.WORKFLOW)
@@ -251,7 +267,6 @@ async def resync_workflows(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
 
     rest_client = create_github_client()
     org_exporter = RestOrganizationExporter(rest_client)
-    port_app_config = cast(GithubPortAppConfig, event.port_app_config)
     config = cast(GithubRepoSearchConfig, event.resource_config)
 
     async for organizations in org_exporter.get_paginated_resources(
@@ -263,10 +278,8 @@ async def resync_workflows(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
             repo_exporter = RestRepositoryExporter(rest_client)
             workflow_exporter = RestWorkflowExporter(rest_client)
 
-            repo_options = ListRepositoryOptions(
-                organization=org_name,
-                type=port_app_config.repository_type,
-                search_params=config.selector.repo_search,
+            repo_options = get_repository_options(
+                org_name, org["type"], config.selector.repo_search
             )
 
             async for repositories in repo_exporter.get_paginated_resources(
@@ -295,8 +308,6 @@ async def resync_workflow_runs(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     repo_exporter = RestRepositoryExporter(rest_client)
     workflow_exporter = RestWorkflowExporter(rest_client)
     workflow_run_exporter = RestWorkflowRunExporter(rest_client)
-
-    port_app_config = cast(GithubPortAppConfig, event.port_app_config)
     config = cast(GithubRepoSearchConfig, event.resource_config)
 
     async for organizations in org_exporter.get_paginated_resources(
@@ -305,10 +316,8 @@ async def resync_workflow_runs(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
         tasks = []
         for org in organizations:
             org_name = org["login"]
-            repo_options = ListRepositoryOptions(
-                organization=org_name,
-                type=port_app_config.repository_type,
-                search_params=config.selector.repo_search,
+            repo_options = get_repository_options(
+                org_name, org["type"], config.selector.repo_search
             )
 
             async for repositories in repo_exporter.get_paginated_resources(
@@ -347,9 +356,6 @@ async def resync_pull_requests(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     org_exporter = RestOrganizationExporter(rest_client)
     repository_exporter = RestRepositoryExporter(rest_client)
     pull_request_exporter = RestPullRequestExporter(rest_client)
-
-    port_app_config = cast(GithubPortAppConfig, event.port_app_config)
-
     config = cast(GithubPullRequestConfig, event.resource_config)
 
     async for organizations in org_exporter.get_paginated_resources(
@@ -358,10 +364,8 @@ async def resync_pull_requests(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
         tasks = []
         for org in organizations:
             org_name = org["login"]
-            repo_options = ListRepositoryOptions(
-                organization=org_name,
-                type=port_app_config.repository_type,
-                search_params=config.selector.repo_search,
+            repo_options = get_repository_options(
+                org_name, org["type"], config.selector.repo_search
             )
 
             async for repos in repository_exporter.get_paginated_resources(
@@ -393,9 +397,6 @@ async def resync_issues(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     org_exporter = RestOrganizationExporter(rest_client)
     repository_exporter = RestRepositoryExporter(rest_client)
     issue_exporter = RestIssueExporter(rest_client)
-
-    port_app_config = cast(GithubPortAppConfig, event.port_app_config)
-
     config = cast(GithubIssueConfig, event.resource_config)
 
     async for organizations in org_exporter.get_paginated_resources(
@@ -404,10 +405,8 @@ async def resync_issues(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
         tasks = []
         for org in organizations:
             org_name = org["login"]
-            repo_options = ListRepositoryOptions(
-                organization=org_name,
-                type=port_app_config.repository_type,
-                search_params=config.selector.repo_search,
+            repo_options = get_repository_options(
+                org_name, org["type"], config.selector.repo_search
             )
 
             async for repos in repository_exporter.get_paginated_resources(
@@ -437,8 +436,6 @@ async def resync_releases(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     org_exporter = RestOrganizationExporter(rest_client)
     repository_exporter = RestRepositoryExporter(rest_client)
     release_exporter = RestReleaseExporter(rest_client)
-
-    port_app_config = cast(GithubPortAppConfig, event.port_app_config)
     config = cast(GithubRepoSearchConfig, event.resource_config)
 
     async for organizations in org_exporter.get_paginated_resources(
@@ -447,10 +444,8 @@ async def resync_releases(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
         tasks = []
         for org in organizations:
             org_name = org["login"]
-            repo_options = ListRepositoryOptions(
-                organization=org_name,
-                type=port_app_config.repository_type,
-                search_params=config.selector.repo_search,
+            repo_options = get_repository_options(
+                org_name, org["type"], config.selector.repo_search
             )
 
             async for repositories in repository_exporter.get_paginated_resources(
@@ -478,8 +473,6 @@ async def resync_tags(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     org_exporter = RestOrganizationExporter(rest_client)
     repository_exporter = RestRepositoryExporter(rest_client)
     tag_exporter = RestTagExporter(rest_client)
-
-    port_app_config = cast(GithubPortAppConfig, event.port_app_config)
     config = cast(GithubRepoSearchConfig, event.resource_config)
 
     async for organizations in org_exporter.get_paginated_resources(
@@ -488,10 +481,8 @@ async def resync_tags(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
         tasks = []
         for org in organizations:
             org_name = org["login"]
-            repo_options = ListRepositoryOptions(
-                organization=org_name,
-                type=port_app_config.repository_type,
-                search_params=config.selector.repo_search,
+            repo_options = get_repository_options(
+                org_name, org["type"], config.selector.repo_search
             )
 
             async for repositories in repository_exporter.get_paginated_resources(
@@ -519,9 +510,6 @@ async def resync_branches(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     org_exporter = RestOrganizationExporter(rest_client)
     repository_exporter = RestRepositoryExporter(rest_client)
     branch_exporter = RestBranchExporter(rest_client)
-
-    port_app_config = cast(GithubPortAppConfig, event.port_app_config)
-
     selector = cast(GithubBranchConfig, event.resource_config).selector
 
     async for organizations in org_exporter.get_paginated_resources(
@@ -530,10 +518,8 @@ async def resync_branches(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
         tasks = []
         for org in organizations:
             org_name = org["login"]
-            repo_options = ListRepositoryOptions(
-                organization=org_name,
-                type=port_app_config.repository_type,
-                search_params=selector.repo_search,
+            repo_options = get_repository_options(
+                org_name, org["type"], selector.repo_search
             )
 
             async for repositories in repository_exporter.get_paginated_resources(
@@ -564,8 +550,6 @@ async def resync_environments(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     org_exporter = RestOrganizationExporter(rest_client)
     repository_exporter = RestRepositoryExporter(rest_client)
     environment_exporter = RestEnvironmentExporter(rest_client)
-
-    port_app_config = cast(GithubPortAppConfig, event.port_app_config)
     config = cast(GithubRepoSearchConfig, event.resource_config)
 
     async for organizations in org_exporter.get_paginated_resources(
@@ -574,11 +558,8 @@ async def resync_environments(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
         tasks = []
         for org in organizations:
             org_name = org["login"]
-
-            repo_options = ListRepositoryOptions(
-                organization=org_name,
-                type=port_app_config.repository_type,
-                search_params=config.selector.repo_search,
+            repo_options = get_repository_options(
+                org_name, org["type"], config.selector.repo_search
             )
 
             async for repositories in repository_exporter.get_paginated_resources(
@@ -607,8 +588,6 @@ async def resync_deployments(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     org_exporter = RestOrganizationExporter(rest_client)
     repository_exporter = RestRepositoryExporter(rest_client)
     deployment_exporter = RestDeploymentExporter(rest_client)
-
-    port_app_config = cast(GithubPortAppConfig, event.port_app_config)
     config = cast(GithubRepoSearchConfig, event.resource_config)
 
     async for organizations in org_exporter.get_paginated_resources(
@@ -617,11 +596,8 @@ async def resync_deployments(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
         tasks = []
         for org in organizations:
             org_name = org["login"]
-
-            repo_options = ListRepositoryOptions(
-                organization=org_name,
-                type=port_app_config.repository_type,
-                search_params=config.selector.repo_search,
+            repo_options = get_repository_options(
+                org_name, org["type"], config.selector.repo_search
             )
 
             async for repositories in repository_exporter.get_paginated_resources(
@@ -650,9 +626,6 @@ async def resync_dependabot_alerts(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     org_exporter = RestOrganizationExporter(rest_client)
     repository_exporter = RestRepositoryExporter(rest_client)
     dependabot_alert_exporter = RestDependabotAlertExporter(rest_client)
-
-    port_app_config = cast(GithubPortAppConfig, event.port_app_config)
-
     config = cast(GithubDependabotAlertConfig, event.resource_config)
 
     async for organizations in org_exporter.get_paginated_resources(
@@ -662,10 +635,8 @@ async def resync_dependabot_alerts(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
         for org in organizations:
             org_name = org["login"]
 
-            repo_options = ListRepositoryOptions(
-                organization=org_name,
-                type=port_app_config.repository_type,
-                search_params=config.selector.repo_search,
+            repo_options = get_repository_options(
+                org_name, org["type"], config.selector.repo_search
             )
 
             async for repositories in repository_exporter.get_paginated_resources(
@@ -695,9 +666,6 @@ async def resync_code_scanning_alerts(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     org_exporter = RestOrganizationExporter(rest_client)
     repository_exporter = RestRepositoryExporter(rest_client)
     code_scanning_alert_exporter = RestCodeScanningAlertExporter(rest_client)
-
-    port_app_config = cast(GithubPortAppConfig, event.port_app_config)
-
     config = cast(GithubCodeScanningAlertConfig, event.resource_config)
 
     async for organizations in org_exporter.get_paginated_resources(
@@ -706,11 +674,8 @@ async def resync_code_scanning_alerts(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
         tasks = []
         for org in organizations:
             org_name = org["login"]
-
-            repo_options = ListRepositoryOptions(
-                organization=org_name,
-                type=port_app_config.repository_type,
-                search_params=config.selector.repo_search,
+            repo_options = get_repository_options(
+                org_name, org["type"], config.selector.repo_search
             )
 
             async for repositories in repository_exporter.get_paginated_resources(
@@ -789,8 +754,6 @@ async def resync_collaborators(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     org_exporter = RestOrganizationExporter(rest_client)
     repository_exporter = RestRepositoryExporter(rest_client)
     collaborator_exporter = RestCollaboratorExporter(rest_client)
-
-    port_app_config = cast(GithubPortAppConfig, event.port_app_config)
     config = cast(GithubRepoSearchConfig, event.resource_config)
 
     async for organizations in org_exporter.get_paginated_resources(
@@ -799,10 +762,8 @@ async def resync_collaborators(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
         tasks = []
         for org in organizations:
             org_name = org["login"]
-            repo_options = ListRepositoryOptions(
-                organization=org_name,
-                type=port_app_config.repository_type,
-                search_params=config.selector.repo_search,
+            repo_options = get_repository_options(
+                org_name, org["type"], config.selector.repo_search
             )
 
             async for repositories in repository_exporter.get_paginated_resources(
@@ -830,9 +791,6 @@ async def resync_secret_scanning_alerts(kind: str) -> ASYNC_GENERATOR_RESYNC_TYP
     org_exporter = RestOrganizationExporter(rest_client)
     repository_exporter = RestRepositoryExporter(rest_client)
     secret_scanning_alert_exporter = RestSecretScanningAlertExporter(rest_client)
-
-    port_app_config = cast(GithubPortAppConfig, event.port_app_config)
-
     config = cast(GithubSecretScanningAlertConfig, event.resource_config)
 
     async for organizations in org_exporter.get_paginated_resources(
@@ -841,10 +799,8 @@ async def resync_secret_scanning_alerts(kind: str) -> ASYNC_GENERATOR_RESYNC_TYP
         tasks = []
         for org in organizations:
             org_name = org["login"]
-            repo_options = ListRepositoryOptions(
-                organization=org_name,
-                type=port_app_config.repository_type,
-                search_params=config.selector.repo_search,
+            repo_options = get_repository_options(
+                org_name, org["type"], config.selector.repo_search
             )
 
             async for repositories in repository_exporter.get_paginated_resources(